+++ bin/fuel-core/src/cli/run/graphql.rs
@@ -12,6 +12,10 @@ pub struct GraphQLArgs {
     #[clap(long = "port", default_value = "4000", env)]
     pub port: u16,
 
+    /// The number of threads to use for the GraphQL service.
+    #[clap(long = "graphql-number-of-threads", default_value = "2", env)]
+    pub graphql_number_of_threads: usize,
+
     /// The size of the batch fetched from the database by GraphQL service.
     #[clap(long = "graphql-database-batch-size", default_value = "100", env)]
     pub database_batch_size: usize,