+++ crates/fuel-core/src/graphql_api/api_service.rs
@@ -58,11 +58,8 @@ use axum::{
     Json,
     Router,
 };
-use fuel_core_metrics::futures::{
-    metered_future::MeteredFuture,
-    FuturesMetrics,
-};
 use fuel_core_services::{
+    AsyncProcessor,
     RunnableService,
     RunnableTask,
     StateWatcher,
@@ -79,6 +76,7 @@ use std::{
         TcpListener,
     },
     pin::Pin,
+    sync::Arc,
 };
 use tokio_stream::StreamExt;
 use tower::limit::ConcurrencyLimitLayer;
@@ -115,6 +113,7 @@ pub struct GraphqlService {
 pub struct ServerParams {
     router: Router,
     listener: TcpListener,
+    number_of_threads: usize,
 }
 
 pub struct Task {
@@ -124,7 +123,7 @@ pub struct Task {
 
 #[derive(Clone)]
 struct ExecutorWithMetrics {
-    metric: FuturesMetrics,
+    processor: Arc<AsyncProcessor>,
 }
 
 impl<F> Executor<F> for ExecutorWithMetrics
@@ -133,9 +132,11 @@ where
     F::Output: Send + 'static,
 {
     fn execute(&self, fut: F) {
-        let future = MeteredFuture::new(fut, self.metric.clone());
+        let result = self.processor.try_spawn(fut);
 
-        tokio::task::spawn(future);
+        if let Err(err) = result {
+            tracing::error!("Failed to spawn a task for GraphQL: {:?}", err);
+        }
     }
 }
 
@@ -159,14 +160,25 @@ impl RunnableService for GraphqlService {
         params: Self::TaskParams,
     ) -> anyhow::Result<Self::Task> {
         let mut state = state.clone();
-        let ServerParams { router, listener } = params;
-        let metric = ExecutorWithMetrics {
-            metric: FuturesMetrics::obtain_futures_metrics("GraphQLFutures"),
+        let ServerParams {
+            router,
+            listener,
+            number_of_threads,
+        } = params;
+
+        let processor = AsyncProcessor::new(
+            "GraphQLFutures",
+            number_of_threads,
+            tokio::sync::Semaphore::MAX_PERMITS,
+        )?;
+
+        let executor = ExecutorWithMetrics {
+            processor: Arc::new(processor),
         };
 
         let server = axum::Server::from_tcp(listener)
             .unwrap()
-            .executor(metric)
+            .executor(executor)
             .serve(router.into_make_service())
             .with_graceful_shutdown(async move {
                 state
@@ -232,6 +244,7 @@ where
     let body_limit = config.config.request_body_bytes_limit;
     let max_queries_resolver_recursive_depth =
         config.config.max_queries_resolver_recursive_depth;
+    let number_of_threads = config.config.number_of_threads;
 
     let schema = schema
         .limit_complexity(config.config.max_queries_complexity)
@@ -296,7 +309,11 @@ where
 
     Ok(Service::new_with_params(
         GraphqlService { bound_address },
-        ServerParams { router, listener },
+        ServerParams {
+            router,
+            listener,
+            number_of_threads,
+        },
     ))
 }